Completed BCD mode fix
I ran the bcd test program and found that it wasn't working in all
cases.  Specifically, input of digits > 9 didn't always work out as they
would on a real 6502.  I was not masking the digits to 4 bits every
place I needed to.  I updated the algorithm as described in the text,
and it all works now, except for the V flag.  That's bugging me, but I
guess it's fine, since that is even less useful than undocumented
opcodes, and I haven't implemented them.