Finished refactoring code log
All console IO is now in em6502.c, rather than the
memory/cpu/instruction files.
Updated operand output so it makes a bit more sense.  Result reported in
log is still variable from one op to another -- it's whatever makes the
most sense to me right now.

Fixed a few small bugs.

Found a spot in the main program option switch where I was unnecessarily
declaring a variable, using it for scanf, and then immediately assigning
the value to a different variable.  Alma caught it becuase the variable
declaration wasn't allowed after a label in the old GCC (apparenly now
legal in C23).  The solution was to correct the input specifier using
%hx, so sscanf new to look for a short int hex value, so it can go right
into the variable.

The IO changes make this a big enough update that I'm calling it version
1.1.